+++ wp-content/plugins/wordpress-beta-tester/src/WPBT/WPBT_Bootstrap.php
@@ -12,20 +12,12 @@
  * WPBT_Bootstrap
  */
 class WPBT_Bootstrap {
-
 	/**
 	 * Holds main plugin file.
 	 *
 	 * @var $file
 	 */
 	protected $file;
-
-	/**
-	 * Holds main plugin directory.
-	 *
-	 * @var $dir
-	 */
-	protected $dir;
 
 	/**
 	 * Holds plugin options.
@@ -37,12 +29,11 @@ class WPBT_Bootstrap {
 	/**
 	 * Constructor.
 	 *
-	 * @param string $file Main plugin file.
+	 * @param  string $file Main plugin file.
 	 * @return void
 	 */
 	public function __construct( $file ) {
 		$this->file = $file;
-		$this->dir  = dirname( $file );
 	}
 
 	/**
@@ -52,19 +43,47 @@ class WPBT_Bootstrap {
 	 */
 	public function run() {
 		$this->deactivate_die_wordpress_develop();
-		$this->load_requires(); // TODO: replace with composer's autoload.
 		$this->load_hooks();
 		self::$options = get_site_option(
 			'wp_beta_tester',
 			array(
-				'stream' => 'point',
-				'revert' => true,
+				'channel'       => 'branch-development',
+				'stream-option' => '',
 			)
 		);
-		// TODO: I really want to do this, but have to wait for PHP 5.4.
-		// TODO: ( new WP_Beta_Tester( $this->file ) )->run( $this->options );
-		$wpbt = new WP_Beta_Tester( $this->file );
-		$wpbt->run( self::$options );
+		$this->v2_v3_settings_migrator();
+
+		( new WP_Beta_Tester( $this->file, self::$options ) )->run();
+	}
+
+	/**
+	 * Gracefully transfer v2 settings to v4.
+	 *
+	 * @since 3.0.0
+	 * @return void
+	 */
+	private function v2_v3_settings_migrator() {
+		if ( isset( self::$options['stream'] ) && ! isset( self::$options['channel'] ) ) {
+			switch ( self::$options['stream'] ) {
+				case 'point':
+					self::$options['channel']       = 'branch-development';
+					self::$options['stream-option'] = '';
+					break;
+				case 'beta-rc-point':
+					self::$options['channel']       = 'branch-development';
+					self::$options['stream-option'] = 'beta';
+					break;
+				case 'unstable':
+					self::$options['channel']       = 'development';
+					self::$options['stream-option'] = '';
+					break;
+				case 'beta-rc-unstable':
+					self::$options['channel']       = 'development';
+					self::$options['stream-option'] = 'beta';
+					break;
+			}
+			update_site_option( 'wp_beta_tester', (array) self::$options );
+		}
 	}
 
 	/**
@@ -94,7 +113,6 @@ class WPBT_Bootstrap {
 		add_action( 'init', array( $this, 'load_textdomain' ) );
 		register_activation_hook( $this->file, array( $this, 'activate' ) );
 		register_deactivation_hook( $this->file, array( $this, 'deactivate' ) );
-		add_filter( 'site_option_wp_beta_tester', array( $this, 'fix_stream' ) );
 	}
 
 	/**
@@ -115,9 +133,7 @@ class WPBT_Bootstrap {
 	 */
 	public function activate() {
 		delete_site_transient( 'update_core' );
-		$wpbt        = new WP_Beta_Tester( $this->file );
-		$wpbt_extras = new WPBT_Extras( $wpbt, self::$options );
-		$wpbt_extras->activate();
+		( new WPBT_Extras( new WP_Beta_Tester( $this->file, self::$options ), self::$options ) )->activate();
 	}
 
 	/**
@@ -129,42 +145,6 @@ class WPBT_Bootstrap {
 	 */
 	public function deactivate() {
 		delete_site_transient( 'update_core' );
-		$wpbt = new WP_Beta_Tester( $this->file );
-		// TODO: ( new WPBT_Extras( $wpbt, self::$options ) )->deactivate();
-		$wpbt_extras = new WPBT_Extras( $wpbt, self::$options );
-		$wpbt_extras->deactivate();
-	}
-
-	/**
-	 * Fix stream option for when `beta-rc` set but current version
-	 * isn't a `beta|RC` version.
-	 *
-	 * @param array $value Array of options values from `wp_beta_tester` option.
-	 *
-	 * @return array
-	 */
-	public function fix_stream( $value ) {
-		if ( is_array( $value ) && 0 === strpos( $value['stream'], 'beta-rc' )
-			&& 1 !== preg_match( '/alpha|beta|RC/', get_bloginfo( 'version' ) ) ) {
-			$value['stream'] = str_replace( 'beta-rc-', '', $value['stream'] );
-		}
-
-		return $value;
-	}
-
-	/**
-	 * <sarcasm>Poor man's autoloader.</sarcasm>
-	 * // TODO: replace with composer's autoload.
-	 *
-	 * @return void
-	 */
-	public function load_requires() {
-		require_once $this->dir . '/src/WPBT/WP_Beta_Tester.php';
-		require_once $this->dir . '/src/WPBT/WPBT_Settings.php';
-		require_once $this->dir . '/src/WPBT/WPBT_Core.php';
-		require_once $this->dir . '/src/WPBT/WPBT_Extras.php';
-		require_once $this->dir . '/src/WPBT/WPBT_Beta_RC.php';
-		require_once $this->dir . '/src/WPBT/WPBT_Help.php';
-		require_once $this->dir . '/vendor/WPConfigTransformer.php';
+		( new WPBT_Extras( new WP_Beta_Tester( $this->file, self::$options ), self::$options ) )->deactivate();
 	}
 }