+++ wp-content/plugins/wordpress-beta-tester/src/WPBT/WPBT_Settings.php
@@ -12,7 +12,6 @@
  * WPBT_Settings
  */
 class WPBT_Settings {
-
 	/**
 	 * Placeholder for main class.
 	 *
@@ -30,8 +29,8 @@ class WPBT_Settings {
 	/**
 	 * Constructor.
 	 *
-	 * @param WP_Beta_Tester $wp_beta_tester Instance of class WP_Beta_Tester.
-	 * @param mixed          $options Saved site options.
+	 * @param  WP_Beta_Tester $wp_beta_tester Instance of class WP_Beta_Tester.
+	 * @param  mixed          $options        Saved site options.
 	 * @return void
 	 */
 	public function __construct( WP_Beta_Tester $wp_beta_tester, $options ) {
@@ -46,18 +45,10 @@ class WPBT_Settings {
 	 */
 	public function run() {
 		$this->load_hooks();
-		// TODO: I really want to do this, but have to wait for PHP 5.4
-		// TODO: ( new WPBT_Core( $this->wp_beta_tester, self::$options ) )->load_hooks();
-		$wpbt_core = new WPBT_Core( $this->wp_beta_tester, self::$options );
-		$wpbt_core->load_hooks();
-		// TODO: ( new WPBT_Extras( $this->wp_beta_tester, self::$options ) )->load_hooks();
-		$wpbt_extras = new WPBT_Extras( $this->wp_beta_tester, self::$options );
-		$wpbt_extras->load_hooks();
-		// TODO: ( new WPBT_Extras( $this->wp_beta_tester, self::$options ) )->skip_autoupdate_email();
-		$wpbt_extras->skip_autoupdate_email();
-		// TODO: ( new WPBT_Help() )->load_hooks();
-		$wpbt_help = new WPBT_Help();
-		$wpbt_help->load_hooks();
+		( new WPBT_Core( $this->wp_beta_tester, self::$options ) )->load_hooks();
+		( new WPBT_Extras( $this->wp_beta_tester, self::$options ) )->load_hooks();
+		( new WPBT_Extras( $this->wp_beta_tester, self::$options ) )->skip_autoupdate_email();
+		( new WPBT_Help() )->load_hooks();
 	}
 
 	/**
@@ -184,7 +175,7 @@ class WPBT_Settings {
 		echo '<nav class="nav-tab-wrapper" aria-label="Secondary menu">';
 		foreach ( $this->settings_tabs() as $key => $name ) {
 			$active = ( $current_tab === $key ) ? 'nav-tab-active' : '';
-			echo( wp_kses_post( '<a class="nav-tab ' . $active . '" href="?page=wp-beta-tester&tab=' . $key . '">' . $name . '</a>' ) );
+			echo wp_kses_post( '<a class="nav-tab ' . $active . '" href="?page=wp-beta-tester&tab=' . $key . '">' . $name . '</a>' );
 		}
 		echo '</nav>';
 	}
@@ -203,7 +194,7 @@ class WPBT_Settings {
 		) {
 			echo '<div class="updated"><p>';
 			esc_html_e( 'Saved.', 'wordpress-beta-tester' );
-			echo( '<span style="padding:0 2em;">' . wp_kses_post( __( 'Why don&#8217;t you <a href="update-core.php">head on over and upgrade now</a>.', 'wordpress-beta-tester' ) ) . '</span>' );
+			echo '<span style="padding:0 2em;">' . wp_kses_post( __( 'Why don&#8217;t you <a href="update-core.php">head on over and upgrade now</a>.', 'wordpress-beta-tester' ) ) . '</span>';
 			echo '</p></div>';
 		}
 	}
@@ -237,7 +228,7 @@ class WPBT_Settings {
 			<h1><?php esc_html_e( 'Beta Testing WordPress', 'wordpress-beta-tester' ); ?></h1>
 			<?php $this->options_tabs(); ?>
 			<div class="updated fade">
-				<p><?php echo( wp_kses_post( __( '<strong>Please note:</strong> Once you have switched your website to one of these beta versions of software, it will not always be possible to downgrade as the database structure may be updated during the development of a major release.', 'wordpress-beta-tester' ) ) ); ?></p>
+				<p><?php echo wp_kses_post( __( '<strong>Please note:</strong> Once you have switched your website to one of these beta versions of software, it will not always be possible to downgrade as the database structure may be updated during the development of a major release.', 'wordpress-beta-tester' ) ); ?></p>
 			</div>
 		<?php