+++ wp-content/plugins/wordpress-beta-tester/CHANGES.md
@@ -1,5 +1,28 @@
 [unreleased]
 
+#### 3.0.6 / 2020-11-21
+* improved flow between _Bleeding edge_ and _Point release_
+
+#### 3.0.5 / 2020-11-18
+* don't show beta as a next version when on RC
+
+#### 3.0.4 / 2020-11-17
+* fix to correctly downgrade from _Bleeding edge_ to _Point release nightlies_.
+* hide stream options other than _Nightlies_ for _Point release_ channel until [new Updates API changes](https://meta.trac.wordpress.org/ticket/5511)
+* add settings for future Updates API above
+* added `channel_settings_migrator()` for switching between `Bleeding edge` and `Point release` channels
+
+#### 3.0.1 - 3.0.3 / 2020-10-27
+* fixed regex to get next versions
+* really didn't need to use `ReflectionClass` 🤦‍♂️, thanks @pbiron
+* use `ReflectionClass` to get static variable `$core_update_constant` from `class WP_Beta_Tester` into `class WPBT_Core`
+
+#### 3.0.0 / 2020-10-23
+* major refactor for new core update API, thanks @dd32!
+* now requires PHP >5.6
+* allows for overrides when using the `WP_AUTO_UPDATE_CORE` constant
+* update on-screen help
+
 #### 2.2.13 / 2020-09-05
 * enclose `WPConfigTransformer` in try/catch